+++ R/jsFunc-html.R
@@ -8,9 +8,11 @@
 #' @param html The HTML/text to place inside the element. Can be either simple
 #' plain text or valid HTML code.
 #' @param add If \code{TRUE}, then append \code{html} to the contents of the element;
-#' otherwise overwrite it. (default: \code{FALSE})
+#' otherwise overwrite it.
 #' @param selector JQuery selector of the elements to target. Ignored if the \code{id}
 #' argument is given.
+#' @param asis If \code{TRUE}, use the ID as-is even when inside a module
+#' (instead of adding the namespace prefix to the ID).
 #' @seealso \code{\link[shinyjs]{useShinyjs}},
 #' \code{\link[shinyjs]{runExample}}
 #' @note \code{shinyjs} must be initialized with a call to \code{useShinyjs()}
@@ -43,8 +45,8 @@
 #'   html(id = "element", add = TRUE, html = input$btn)
 #' }
 #' @export
-html <- function(id, html, add, selector) {
+html <- function(id = NULL, html = NULL, add = FALSE, selector = NULL, asis = FALSE) {
   fxn <- "html"
-  params <- as.list(match.call())[-1]
+  params <- list(id = id, html = html, add = add, selector = selector, asis = asis)
   jsFuncHelper(fxn, params)
 }