+++ R/jsFunc-stateFuncs.R
@@ -13,10 +13,13 @@
 #' then you can use \code{toggleState} instead: \code{toggleState(id, test())}.
 #'
 #' @param id The id of the input element/Shiny tag
-#' @param condition An optional argument to \code{toggleState}, see 'Details' below.
+#' @param condition An optional argument to \code{toggleState}. The element will
+#' be enabled when the \code{condition} is \code{TRUE}, and disabled otherwise.
 #' @param selector Query selector of the elements to target.  Ignored if the \code{id}
 #' argument is given. For example, to disable all text inputs, use
 #' \code{selector = "input[type='text']"}
+#' @param asis If \code{TRUE}, use the ID as-is even when inside a module
+#' (instead of adding the namespace prefix to the ID).
 #' @seealso \code{\link[shinyjs]{useShinyjs}},
 #' \code{\link[shinyjs]{runExample}}
 #' \code{\link[shinyjs]{disabled}}
@@ -75,24 +78,24 @@ NULL
 
 #' @export
 #' @rdname stateFuncs
-enable <- function(id, selector) {
+enable <- function(id = NULL, selector = NULL, asis = FALSE) {
   fxn <- "enable"
-  params <- as.list(match.call())[-1]
+  params <- list(id = id, selector = selector, asis = asis)
   jsFuncHelper(fxn, params)
 }
 
 #' @export
 #' @rdname stateFuncs
-disable <- function(id, selector) {
+disable <- function(id = NULL, selector = NULL, asis = FALSE) {
   fxn <- "disable"
-  params <- as.list(match.call())[-1]
+  params <- list(id = id, selector = selector, asis = asis)
   jsFuncHelper(fxn, params)
 }
 
 #' @export
 #' @rdname stateFuncs
-toggleState <- function(id, condition, selector) {
+toggleState <- function(id = NULL, condition = NULL, selector = NULL, asis = FALSE) {
   fxn <- "toggleState"
-  params <- as.list(match.call())[-1]
+  params <- list(id = id, condition = condition, selector = selector, asis = asis)
   jsFuncHelper(fxn, params)
 }